# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Posts API Laravel",
        "version": "0.1"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "description": "Register a new user",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "description": "User object that needs to be registered",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "yago camara"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "yago@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                              "schema": {
                                "type": "object",
                                "properties": {
                                  "token": {
                                    "type": "string",
                                    "description": "The user token."
                                  }
                                }
                              }
                            }
                          }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate an existing user",
                "description": "Authenticate an existing user",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User authenticated successfully",
                        "content": {
                            "application/json": {
                              "schema": {
                                "type": "object",
                                "properties": {
                                  "access_token": {
                                    "type": "string",
                                    "description": "The user token."
                                  }
                                }
                              }
                            }
                          }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Lista todos os posts ou filtra por tag",
                "operationId": "e9eaf7df702a94f4443677bd1a36cfe6",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Filtrar por tag",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de posts retornada com sucesso",
                        "content": {
                            "application/json": {
                              "schema": {
                                "$ref": "#/components/schemas/Post"
                              }
                            }
                          }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Cria uma nova postagem",
                "description": "Store a newly created resource in storage.",
                "operationId": "4a657e51a0066c64781851f2fd2651e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 150,
                                        "example": "Título da postagem"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Conteúdo da postagem"
                                    },
                                    "author": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Autor da postagem"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "tag1",
                                            "tag2"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Postagem criada com sucesso",
                        "content": {
                            "application/json": {
                              "schema": {
                                "$ref": "#/components/schemas/Post"
                              }
                            }
                          }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Exibe um post específico",
                "operationId": "b0f9d92879218f68a376b6edc6467c67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post retornado com sucesso",
                        "content": {
                            "application/json": {
                              "schema": {
                                "$ref": "#/components/schemas/Post"
                              }
                            }
                          }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
               
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Atualiza um post existente",
                "operationId": "929614537d76a081a92170a41ed3bdb3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post atualizado com sucesso",
                        "content": {
                            "application/json": {
                              "schema": {
                                "$ref": "#/components/schemas/Post"
                              }
                            }
                          }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Exclui um post existente",
                "operationId": "fad1726bd24cd459fc1525b67985df39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post excluído com sucesso" 
                       
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "tags": [
        {
            "name": "Authentication",
            "description": "Endpoints for user authentication"
        }
    ],
    "components": {
        "schemas":{
          "Post":{
            "type": "object",
            "properties":{
                "id":{
                    "type":"integer",
                    "description": "Post ID"
                },
                "title":{
                    "type": "string",
                    "description":"Post title"
                },
                "author":{
                    "type":"string",
                    "description":"Post's author"
                },
                "content":{
                    "type":"string",
                    "description":"Post's content"
                },
                "tags":{
                    "type":"array",
                    "description":"Post's tags"
                }
            }
         }
        },
        "securitySchemes": {
          "bearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
          }
        }
      }
    
}